backend/app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
databases
openai
python-dotenv
pandas